add distance on dc call from donor

+++ src/main/java/com/easyhelp/application/model/dto/dcp/outgoing/DCPDonationRequestDetailsDTO.java
@@ -20,7 +20,7 @@ public class DCPDonationRequestDetailsDTO extends BaseOutgoingDTO {
     private Double quantity;
     private RequestUrgency urgency;
     private RequestStatus status;
-    private Integer distance;
+    private Double distance;
     private Boolean hasCommitted;
 
     public DCPDonationRequestDetailsDTO(DonationRequest donationRequest, DonationCenter donationCenter) {
